backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "AI Readiness Decision Intelligence System — Backend"
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.1",
    "pydantic>=2.7.1",
    "sqlalchemy>=2.0.30",
    "boto3>=1.34.131",
    "python-dotenv>=1.0.1",
    "httpx>=0.27.0",
    "reportlab>=4.2.0",
    "pytest>=9.0.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
